+++ lib/Skipper.py
@@ -2,12 +2,11 @@ from lib.PathFunctions import PathFunction
 
 class Skip:
     def __init__(self):
-        self.exist = False 
+        self.PathFunctions = PathFunction()
         self.path_list = []
         self.netloc_list = []
         self.parameter_list = {}
         self.uparameter_list = {}
-        self.path_fn = PathFunction()
 
     def add_path(self, path_to_add: str) -> bool:
         if path_to_add in self.path_list:
@@ -17,7 +16,7 @@ class Skip:
             return True
 
     def add_parameter(self, url: str, parameter_list: list) -> list:
-        url = self.path_fn.ender(url, '?')
+        url = self.PathFunctions.ender(url, '?')
         if url in self.parameter_list:
              var = self.parameter_list[url]
              var.update(set(parameter_list))
@@ -41,26 +40,16 @@ class Skip:
             self.netloc_list.append(netloc)
             return True
 
-    def check_path(self, path_to_add: str) -> bool: #return True if found
-        if path_to_add in self.path_list:
-            return True
-        else:
-            return False
+    def check_path(self, path: str) -> bool:
+        return bool(path in self.path_list)
 
-    def check_netloc(self, netloc: str) -> bool: #return True if found
-        if netloc in self.netloc_list:
-            return True
-        else:
-            return False
+    def check_netloc(self, netloc: str) -> bool:
+        return bool(netloc in self.netloc_list)
 
-    def check_parameter(self, url: str, parameter: str) -> bool: #return True if found
-        url = self.path_fn.ender(url, '?')
-        try:
-            if self.parameter_list[url]:
-                self.exist = True
-        except:
-            self.exist = False
-        if self.exist:
+    def check_parameter(self, url: str, parameter: str) -> bool:
+        url = self.PathFunctions.ender(url, '?')
+        exist = bool(self.parameter_list.get(url))
+        if exist:
             for self_parameter in self.parameter_list[url]:
                 if self_parameter == parameter:
                     return True